refactor(cli): extract interactive terminal lifecycle
